\name{SeqFeatR}
\alias{SeqFeatR-package}
\docType{package}
\title{Associates point and tuple mutations with sequence features.}
\description{Package includes functions to analyze sequence alignments
  in FASTA format with known HLA types or other features. In case of HLA
  as a feature, possible HLA epitopes are predicted based on clusters of
  HLA-associated mutations. Various plots are provided to visualize results.}
\details{
\tabular{ll}{Package: \tab SeqFeatR\cr
Type: \tab Package\cr
Title: \tab Sequence-feature associations\cr
Version: \tab 0.1.7\cr
Date: \tab 2015-02-10\cr
Author: \tab Bettina Budeus\cr
Maintainer: \tab Bettina Budeus <bettina.budeus@stud.uni-due.de>\cr
Depends: \tab R(>=3.1.2), tcltk, relimp, tcltk2, Biostrings, plyr, plotrix, phangorn,\cr
 qvalue, widgetTools, ggplot\cr
License: \tab GPL (>= 3)\cr}
}
\author{Bettina Budeus}

\keyword{ package }


