\name{SeqFeatRGUI}
\alias{SeqFeatR_GUI}
\title{GUI for SeqFeatR}
\description{A GUI for all program features of SeqFeatR}
\usage{SeqFeatR_GUI()}

\details{This is a GUI for the program SeqFeatR. The first tab of the
  GUI deals with the association of alignment positions with features
  (e.g. HLA types); a subset of the functionality is provided by a web-server
  at http://seqfeatr.zmb.uni-due. See also video tutorial at 
  https://www.youtube.com/watch?v=-CYidGPE6dw and vignette in this package. }

\author{Bettina Budeus}



\examples{
\dontrun{SeqFeatR_GUI()
}
}
