\name{SeqFeatR}
\alias{SeqFeatR-package}
\docType{package}
\title{Associates point and tuple mutations with sequence features.}
\description{Package includes functions to analyse sequence data in FASTA Format with known HLA types or other features and calculates in the case of HLA as feature possible epitopes from them. Further it calculates possible co-mutations from this sequences and generates graphical output in form of a pdfs.}
\details{
\tabular{ll}{Package: \tab SeqFeatR\cr
Type: \tab Package\cr
Title: \tab Calculates possible epitopes and co-mutations\cr
Version: \tab 1.5\cr
Date: \tab 2014-06-04\cr
Author: \tab Bettina Budeus\cr
Maintainer: \tab Bettina Budeus <rokkaku-fight@gmx.de>\cr
Depends: \tab R(>=3.1.0), tcltk, relimp, tcltk2, Biostrings, plyr, plotrix, phangorn,\cr
 qvalue, widgetTools\cr
License: \tab GPL (>= 2)\cr}
}
\author{Bettina Budeus}

\keyword{ package }


